Do not set default crush rule when creating pools
A recent change to microceph *1 caused the default crush rule called
`replicated_rule` to be removed. Instead a rule called
`microceph_auto_osd` is created on single node deployments. If/when
the cluster is expanded sufficently a new rule called
`microceph_auto_host` is created automatically and existing pools
migrated to use this, all this is done automatically be microceph.

This change stop a profile being set at all to take advantage of this
new behaviour.

*1 https://github.com/canonical/microceph/commit/14dfe4513fd4c2fb80f7d24b5f4aabb84710d326