Implement additional broker operations (#27)
* Implement additional broker operations

This PR implements most of the broker operations missing in
charm-microceph (compared to what's in charms.ceph). The only
exceptions so far are those related to cache-tiering (because
they need additional classes from charmhelpers). Additionally,
most of the error checking stuff has been removed and delegated
to the external calls. Note that most of the operations that
call out to 'radosgw-admin' will not work unless RadosGW is
installed, but they can be placeholders until we know what
to do with these.

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>

* Add unit tests for the broker

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>

* Test replicas number as well.

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>

* Fix request access to 'name' key

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>

* Use a jump table and test a few more ops

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>

* Style fixes.

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>

---------

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>