Action add-osd: Add all disks irrespective of errors
Currently add-osd tries to add disk one by one and
the action execution returns failure on the first
error encountered without trying to add rest of the
disks.
However it is desirable to add all the disks specified
by user and returns the aggregated output for each disk
in case of success and failure.