Add config option enable-rgw
Add new option enable-rgw.
If the value is "*", enable the rgw on all
microceph nodes.
If the value is "", disable the rgw on all
microceph nodes.

Add cluster client to interact with microceph
daemon over unix socket. For now, add a function
to get all services from the cluster. In future,
this client can be enhanced to bootstrap the
cluster, join new nodes etc.