Delete rgw config only if it exists in db
Currently microceph daemon API is called to
delete rgw config irrespective of config exists
or not as the daemon call does not return
error if config does not exist in db. However
this causes time delays in running ceph command
for each config given there are almost 15 configs
introduced for rgw.
Add a check to call delete config API call
only when the config exists in db.

Signed-off-by: Hemanth Nakkina <hemanth.nakkina@canonical.com>