Fix getting handler of identity service
Getting the handler for the identity service involves a call to
'secret-get', which can fail if the relation endpoint is gone. Normally,
this would be detected by the usual mechanisms of the ops framework.
However, for cross-model relations, the charm may still want to fetch
it, even if the relation is broken. As such, this patchset wraps the
getting part in a try/except to prevent this error from arising.

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>