Add host-only param to list-disks action (#147)
Currently list-disks action returns OSDs on
the cluster and unpartitioned disks on the
host. Microceph snap supports --host-only
flag to disks list command to list OSDs on
the cluster [1]

Add param to list-disks action host-only.
If this param is set to True, list only the
OSD disks on the host instead of on the cluster.

Related Bug: https://bugs.launchpad.net/snap-openstack/+bug/2098626

[1] https://github.com/canonical/microceph/commit/5a944d53179a92f2c7aa851fca77199bed8ca5fb