Improve upgrade robustness (#149)
In the upgrade process we synchronize by checking if the cluster is
healthy after initiating an upgrade, with the expectation that an
upgrade will restart services and we need to wait before services are
up again.

However, there is a race in there, as we might check health even
before services have restarted, or service status has propagated.

This patch changes the process such that we require successful
consecutive health checks with a waiting period in between before we
continue with the upgrade process to avoid this race.

Signed-off-by: Peter Sabaini <peter.sabaini@canonical.com>