Feat: ceph-admin interface for adopting an external ceph cluster (#212)
This change adds an adopt integration which allows the charm  to bootstrap the microceph cluster using
an existing charmed ceph cluster. To do this, we need to instruct the charm to not bootstrap a new ceph
cluster when the leader is first up, this is signaled using a charm config `wait-for-adoption` basis which
the charm waits indefinitely until bootstrap data is made available on the adopt-ceph relation.
---------

Signed-off-by: Utkarsh Bhatt <utkarsh.bhatt@canonical.com>